v1/requirements.txt
langchain-community
langchain
langchain-openai
langchain-experimental
azure-storage-blob
neo4j
pypdf